Docker support added
