+++ Project-EPA-OpenAI/EPA-WebAPI/Dockerfile
@@ -6,20 +6,19 @@
 FROM mcr.microsoft.com/dotnet/aspnet:7.0 AS base
 WORKDIR /app
 EXPOSE 80
-EXPOSE 443
 
 FROM mcr.microsoft.com/dotnet/sdk:7.0 AS build
 WORKDIR /src
-COPY ["Project-EPA.csproj", "."]
-RUN dotnet restore "./Project-EPA.csproj"
+COPY ["EPA-WebAPI/EPA-WebAPI.csproj", "EPA-WebAPI/"]
+RUN dotnet restore "EPA-WebAPI/EPA-WebAPI.csproj"
 COPY . .
-WORKDIR "/src/."
-RUN dotnet build "Project-EPA.csproj" -c Release -o /app/build
+WORKDIR "/src/EPA-WebAPI"
+RUN dotnet build "EPA-WebAPI.csproj" -c Release -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "Project-EPA.csproj" -c Release -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "EPA-WebAPI.csproj" -c Release -o /app/publish /p:UseAppHost=false
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "Project-EPA.dll"]
+ENTRYPOINT ["dotnet", "EPA-WebAPI.dll"]
