+++ Project-EPA-OpenAI/Project-EPA-OpenAI.sln
@@ -1,0 +1,38 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33414.496
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Microservices", "Microservices", "{BD61505E-E66F-4F19-90D0-679BCFA9A75B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "OpenWebApi", "OpenWebApi", "{A86F8684-B2CB-4090-9AA5-6E46AAB29A2A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EPA-WebAPI", "EPA-WebAPI\EPA-WebAPI.csproj", "{55A8709F-F778-4B7B-A757-33A4A549A0C9}"
+EndProject
+Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{08A37222-BC7C-4D2E-A508-375282553C34}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{55A8709F-F778-4B7B-A757-33A4A549A0C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55A8709F-F778-4B7B-A757-33A4A549A0C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55A8709F-F778-4B7B-A757-33A4A549A0C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55A8709F-F778-4B7B-A757-33A4A549A0C9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{08A37222-BC7C-4D2E-A508-375282553C34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{08A37222-BC7C-4D2E-A508-375282553C34}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{08A37222-BC7C-4D2E-A508-375282553C34}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{08A37222-BC7C-4D2E-A508-375282553C34}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{55A8709F-F778-4B7B-A757-33A4A549A0C9} = {A86F8684-B2CB-4090-9AA5-6E46AAB29A2A}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EE39D7F9-23B4-442F-9727-B34058901D3E}
+	EndGlobalSection
+EndGlobal
