+++ Project-EPA-OpenAI/docker-compose.override.yml
@@ -1,0 +1,8 @@
+version: '3.4'
+
+services:
+  epa-webapi:
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Development
+    ports:
+      - "80"
