+++ Project-EPA-OpenAI/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "Docker Compose": {
+      "commandName": "DockerCompose",
+      "commandVersion": "1.0",
+      "serviceActions": {
+        "epa-webapi": "StartDebugging"
+      }
+    }
+  }
+}
