--- Project-EPA/Project-EPA.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33414.496
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Project-EPA", "Project-EPA.csproj", "{7A39C0FD-EDCC-45B1-864D-1FB458BA36C7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7A39C0FD-EDCC-45B1-864D-1FB458BA36C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7A39C0FD-EDCC-45B1-864D-1FB458BA36C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7A39C0FD-EDCC-45B1-864D-1FB458BA36C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7A39C0FD-EDCC-45B1-864D-1FB458BA36C7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {65A60F05-6930-47C8-961F-6009A1AAD22C}
-	EndGlobalSection
-EndGlobal
